changes for verification and authorization

See merge request TejShrestha/laravel_intern_practice!12

+++ app/Http/Controllers/ProductController.php
@@ -5,8 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
 class ProductController extends Controller
 {
+    private function isAdmin()
+    {
+        $user = Auth::user();
+        return $user->roles->contains('name', 'admin');
+    }
     public function index(Request $request)
     {
         $category_id = $request->input("category_id");
@@ -17,6 +23,9 @@ class ProductController extends Controller
 
     public function create(Product $product)
     {
+        if (!$this->isAdmin()) {
+            return redirect()->back()->with('error', 'Access denied. Admins only.');
+        }
         $categories = Category::all();
         return view('product.create', compact('categories'));
     }
@@ -105,6 +114,12 @@ class ProductController extends Controller
 
     public function destroy(Product $product)
     {
+        if (!$this->isAdmin()) {
+            return redirect()->back()->with('error', 'Access denied. Admins only.');
+        }
+        if ($product->image_path && file_exists(public_path($product->image_path))) {
+            unlink(public_path($product->image_path));
+        }
         $product->delete();
         return redirect('/category')->with('success','Product Deleted Successfully');
     }
